+++ tests/testthat/bad_missing/R/func.R
@@ -1,0 +1,4 @@
+my_func <- function(x, y) {
+  if (missing(x)) x <- 1
+  x + y
+}